+++ tests/tests_unit/graph/test_loaders/test_dms_loader.py
@@ -63,7 +63,7 @@ def test_imf_attribute_nodes():
 
     dms_rules = InformationToDMS().transform(info_rules)
 
-    store = NeatGraphStore.from_oxi_store()
+    store = NeatGraphStore.from_oxi_local_store()
     store.add_rules(info_rules)
     store.write(RdfFileExtractor(IMF_EXAMPLE))