+++ tests/tests_unit/rules/test_importers/test_inference_importer.py
@@ -15,7 +15,7 @@ from tests.data import car
 
 
 def test_rdf_inference():
-    store = NeatGraphStore.from_oxi_store()
+    store = NeatGraphStore.from_oxi_local_store()
     extractor = RdfFileExtractor(nordic44_knowledge_graph, base_uri="http://nordic44.com/")
     store.write(extractor)
 
@@ -53,7 +53,7 @@ def test_rdf_inference_with_removal_of_unknown_type():
     EX = Namespace("http://example.org/")
     SUBSTATION = Namespace("http://example.org/substation/")
     TERMINAL = Namespace("http://example.org/terminal/")
-    store = NeatGraphStore.from_oxi_store()
+    store = NeatGraphStore.from_oxi_local_store()
 
     store.dataset.add((EX.substation1, RDF.type, SUBSTATION.Substation))
     store.dataset.add((EX.substation2, RDF.type, SUBSTATION.Substation))
@@ -74,7 +74,7 @@ def test_rdf_inference_with_removal_of_unknown_type():
 
 
 def test_rdf_inference_with_none_existing_node():
-    store = NeatGraphStore.from_oxi_store()
+    store = NeatGraphStore.from_oxi_local_store()
     extractor = RdfFileExtractor(DATA_FOLDER / "low-quality-graph.ttl")
     store.write(extractor)