+++ docs/CHANGELOG.md
@@ -27,6 +27,7 @@ Changes are grouped as follows:
 ### Added
 - Fast serialization of NeatGraphStore
 - `neat.to.session(...)` and `neat.from.session(...)` methods to save and load a `NeatSession` object to and from a file.
+- Internal support for remote oxigraph store
 - Neat can now read a knowledge graph from DMS with the `neat.read.cdf.graph(...)` method.
 
 ### Improved